refactor: move 'auth' module into 'infra' directory
