+++ src/common/guards/roles.guard.ts
@@ -8,7 +8,7 @@ import { Reflector } from "@nestjs/core";
 import { PrismaService } from "src/infra/prisma/prisma.service";
 import { Role } from "../enums/role.enum";
 import { ROLES_KEY } from "../decorators/roles.decorator";
-import { IRequestWithUser } from "src/auth/interfaces/request-user.interface";
+import { IRequestWithUser } from "src/infra/auth/interfaces/request-user.interface";
 
 @Injectable()
 export class RolesGuard implements CanActivate {
