+++ src/infra/auth/interfaces/jwt-payload.interface.ts
