+++ src/infra/auth/strategies/jwt.strategy.ts
