+++ src/infra/auth/v1/dto/swagger/auth-api-response.dto.ts
