+++ test/helpers/create-test-app.helper.ts
@@ -1,7 +1,7 @@
 import { ConfigModule, ConfigService } from "@nestjs/config";
 import { Test, TestingModule } from "@nestjs/testing";
 import { AdminModule } from "src/admin/admin.module";
-import { AuthModule } from "src/auth/auth.module";
+import { AuthModule } from "src/infra/auth/auth.module";
 import { PrismaModule } from "src/infra/prisma/prisma.module";
 import { PrismaService } from "src/infra/prisma/prisma.service";
 import { UserModule } from "src/domain/user/user.module";
