+++ test/helpers/test-auth.helper.ts
@@ -2,8 +2,8 @@
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 
 import { Body, INestApplication } from "@nestjs/common";
-import { ILogin } from "src/auth/interfaces/login.interface";
-import { AuthApiResponseDto } from "src/auth/v1/dto/swagger/auth-api-response.dto";
+import { ILogin } from "src/infra/auth/interfaces/login.interface";
+import { AuthApiResponseDto } from "src/infra/auth/v1/dto/swagger/auth-api-response.dto";
 import request from "supertest";
 
 export class TestAuthHelper {
